#111 QuotesViewModel + Enum for Notification time
