+++ QuoteDroplet/Views/Helpers/NotificationTimeEnum.swift
@@ -1,0 +1,13 @@
+//
+//  NotificationTimeEnum.swift
+//  Quote Droplet
+//
+//  Created by Daniel Agapov on 2024-10-24.
+//
+
+import Foundation
+
+enum NotificationTime {
+    case previouslySelected
+    case defaultScheduled
+}
